backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
opencv-python>=4.10.0
mediapipe>=0.10.0
openai-whisper>=20231117
torch>=2.0.0
numpy<2.0.0
transformers>=4.40.0
openai>=1.0.0
anthropic>=0.30.0
python-dotenv>=1.0.0
Pillow>=10.0.0
moviepy>=1.0.3
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic-settings>=2.0.0
